Improve cart collector
