--- src/Sylius/Bundle/CoreBundle/Resources/views/Collector/debug_cart.html.twig
@@ -1,98 +1,0 @@
-{% extends '@WebProfiler/Profiler/layout.html.twig' %}
-
-{% block toolbar %}
-    {% import "@SyliusShop/Common/Macro/money.html.twig" as money %}
-
-    {% set icon %}
-        <span class="sf-toolbar-value">Cart ({{ collector.quantity|default(0) }})</span>
-    {% endset %}
-    {% set text %}
-        <div class="sf-toolbar-info-group">
-            <div class="sf-toolbar-info-piece">
-                <b>ID</b>
-                <span>{{ collector.cartId }}</span>
-            </div>
-            <div class="sf-toolbar-info-piece">
-                <b>Subtotal</b>
-                <span>{{ money.convertAndFormat(collector.subtotal, collector.currency) }}</span>
-            </div>
-            <div class="sf-toolbar-info-piece">
-                <b>Total</b>
-                <span>{{ money.convertAndFormat(collector.total, collector.currency) }}</span>
-            </div>
-        </div>
-        <div class="sf-toolbar-info-group">
-            <div class="sf-toolbar-info-piece">
-                <span><a href="{{ profiler_url ~ '?panel=sylius_core_debug_cart' }}" rel="help">View details</a></span>
-            </div>
-        </div>
-    {% endset %}
-
-    {% include '@WebProfiler/Profiler/toolbar_item.html.twig' with {'link': profiler_url} %}
-{% endblock %}
-
-{% block menu %}
-<span class="label">
-    <span class="icon">{{ include('@WebProfiler/Icon/form.svg') }}</span>
-    <strong>Cart</strong>
-</span>
-{% endblock %}
-
-{% block panel %}
-
-{% import "@SyliusShop/Common/Macro/money.html.twig" as money %}
-
-<h2>Cart</h2>
-    <table>
-        <tr>
-            <th>ID</th>
-            <th>Currency</th>
-            <th>Locale</th>
-            <th>Subtotal</th>
-            <th>Total</th>
-        </tr>
-        <tr>
-            <td>{{ collector.cartId }}</td>
-            <td>{{ collector.currency }}</td>
-            <td>{{ collector.locale }}</td>
-            <td>{{ money.convertAndFormat(collector.subtotal, collector.currency) }}</td>
-            <td>{{ money.convertAndFormat(collector.total, collector.currency) }}</td>
-        </tr>
-    </table>
-
-<h2>States</h2>
-    <table>
-        <tr>
-            <th>Main</th>
-            <th>Checkout</th>
-            <th>Shipping</th>
-            <th>Payment</th>
-        </tr>
-        <tr>
-            <td>{{ collector.states.main }}</td>
-            <td>{{ collector.states.checkout }}</td>
-            <td>{{ collector.states.shipping }}</td>
-            <td>{{ collector.states.payment }}</td>
-        </tr>
-    </table>
-
-{% if collector.items|length > 0 %}
-<h2>Items</h2>
-    <table>
-        <tr>
-            <th>ID</th>
-            <th>Product (ID)</th>
-            <th>Variant (ID / Code)</th>
-            <th>Quantity</th>
-        </tr>
-        {% for item in collector.items %}
-        <tr>
-            <td>{{ item.id }}</td>
-            <td>{{ item.productName }} ({{item.productId}})</td>
-            <td>{{ item.variantName }} ({{ item.variantId }} / {{ item.variantCode }})</td>
-            <td>{{ item.quantity }}</td>
-        </tr>
-        {% endfor %}
-    </table>
-{% endif %}
-{% endblock %}
